requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
python-dotenv
pypdf
unstructured
python-docx
pandas
langchain-community
